--- Sources/Router/router.swift
@@ -1,40 +1,0 @@
-import Foundation
-import Quick
-import Nimble
-import Router
-import Requests
-
-class RouterSpec: QuickSpec {
-    override func spec() {
-        describe("#Router") {
-            var router = Router!
-
-            beforeEach {
-                router = Router()
-            }
-
-            it ("return a 200 OK response if the method/url combo is correct") {
-                let validRequest = HttpRequest(
-                        method: "GET",
-                        url: "/",
-                        version: "HTTP/1.1",
-                        headers: [
-                            "Host": "localhost:5000",
-                            "User-Agent": "curl/7.54.0",
-                            "Accept": "*/*"
-                        ]
-                )
-
-                let responseOK = HttpResponse(
-
-                )
-                let response = router.checkRoute(request: validRequest)
-                expect(response).to(equal(responseOK))
-            }
-
-            it ("returns a 400 NotFound if the method/url combo is not correct") {
-
-            }
-        }
-    }
-}
